DLS-1851 - Changed cost in SBA01 to an Int

+++ src/main/scala/uk/gov/hmrc/ct/computations/SBA01.scala
@@ -40,6 +40,6 @@ case class Building(
                      postcode: Option[String],
                      earliestWrittenContract: Option[LocalDate],
                      nonResidentialActivityStart: Option[LocalDate],
-                     cost: Option[BigDecimal],
+                     cost: Option[Int],
                      claim: Option[Int]
                    )
